End-to-end wake call in dev server
This adds a lot more of the overall shape of the application, and
fleshes out the dev-server enough to allow a working "wake" API call
(checked using `wscat`).

The main (API) jobs ahead are:
1. Implement DynamoDB-backed persistence
2. Test and write the game logic
3. Get it running in AWS

The last of these points depends on CDK adding support for
Websockets. At some point I'll give up waiting and do it in
CloudFormation instead.