Change DB types/logic to support  Join Game
Planning JoinGame was a reminder that we need to be able to look up
games by the game code, to allow player's to join. Minor refactoring
of the GameDB type and the logic around it enabled this use case.

Some sketches for JoinGame are here, as is a stub for Start Game,
which is probably better as its own API call. Multiple simple API
calls are preferred to a single complex one.