Adds creator to Game model, tests for player order
Integration tests now support player order. Including the creator on
the game was a bit of a refactor, but allows us to enforce some
operations (e.g. start game) are only done by the creator. This
preserves the property that the game object is only writable by a
single player at a time to help eliminate race conditions.