Stub of bid and supporting refactor
The bid implementation highlighted that the target amount was not
carried over to the bidding round's data. This is now included in the
Game representations.

Also refactors the bid / passed properties at the same time. These did
not need to be wrapped in Options, an indirection that only added
unnecessary casese.

Fixes and adds some test cases to catch related edge cases and bugs.