Add disc removal and player elimination
Discs are removed as punishment for flipping a Thorn. When a player
has had all discs removed they are eliminated. This menas they do not
take part in placing and cannot make bids.

Both of these mechanisms ar implemented, although some additional
testing is required on previously built endpoints to make sure it is
behaving properly (and to prevent regressions).

The GameSummary type that is sent to clients now also includes the
number of discs each player possesses.As well as being useful
strategic information, this is essential for spotting players that
have been eliminated.