Bug fixes and improvements to flipping
Testing flip revealed that the round's target was lost when the player
got saved. This is because the player bids were not carried over into
the flipping round, which caused those values to not be persisted.

The flipping round now also carries the bids. This fixes the
persistence problem, but can also provide tactical information to the
players. It's now also carried over to the round summary information
that gets sent to the client.