Hello world frontend app with datastructures modelled in Elm
The modelling highlighted that the round summaries need type tags so
they can be decoded properly. The API is updated to accomodate this,
but I don't love he current approach. I may switch to doing this at
the codec level in the future.

This is a Hello World UI with no functionality, but is wired up in
roughly the structure we'll be using for the app.