Wire prototype InitialDiscs UI, refactor update
The update function is now split up a bit to make it easier to
navigate in IntelliJ.  The basic server message behaviour is in place
now for all messages, without any UX polish for each screen.

The PlaceDisc screen now works, which means the game can progress as
far as the first player's disc/bid in the web frontend.