Vaguely playable prototype frontend
Web frontend is end-to-end enough to be playable. Passing, flipping
and new rounds implemented, as well as the round end display.

There's no layout or style yet, and this is only playable if the
player keeps a lot of game state in their head.

Some API bugs were uncovered during this work, they are also fixed.