Some UI improvements and WATs
UI improvements were tedious, so I added some Web Automation Testing
with the goal of exporting the Elm Debugger Log for use
in-browser. This doesn't seem to be possible at the moment, but as a
side-effect there is a prototype of WAT for Thorn which will be useful
for pre and post release testing.

These WATs produce screenshots as an output, I've checked these files
into the repository and may use them from the README in he future if
they are kept broadly up to date.

Some UI improvements as well.