Work on AWS deployment
Cloudformation templates exist and work, but the application doesn't
yet successfully send messages back to the client.

As well as the infrastrucure, this fills out the Lambda project, with
'production' instances of db and messaging (the latter is not yet
working).

This also introduces both sbt-assembly and sbt-native-packager. I've
been experimenting with both because the jar is currently far too
large for a Lambda Function.