Refactor WelcomeMessage and Library
Welcome messages now contain the player and game names so that all the
information needed to do "rejoins" is in one place. The library can
now be a list of welcome messages, instead of the more complicated
dictionary type that was previously required. This simplifies the
game's lifecycle, and will make it much easier to save the library to
persistent storage in the browser.