Persist games to local storage so they can be resumed
When a welcome message is received, it goes into Local Storage. A
welcome message is all the information the UI needs to re-enter a
game, so this allows players to resume a game after closing the tab
(or accidentally reloading).

Adds a bit of JS to manage the local storage aspects and some ports to
go between.